v0.80 re-upload
Github automatically changed the line endings from CRLF to LF without notification. That's the default setting for new repos. Too bad that Windows batches come in CRLF AND MIGHT BREAK IF YOU CONVERT THEM TO LF.
Modified the attributes for the repository and re-uploaded the files with correct line endings.
Huge thanks to Fox2k11 for reporting that. I hate Github.